chore: update dependencies and refactor test utilities
- Bump esbuild version to 0.28.0 and add vite as a dependency.
- Refactor test utilities by moving functions from `@local-save/test/test-utils` to `@local-save/test/setup`.
- Remove the old test-utils file as its functions are now centralized in setup.
- Update integration and unit tests to import from the new setup file.
- Adjust vitest configuration to conditionally set the silent option based on test log settings.
