+++ package.json
@@ -59,7 +59,7 @@
         "@vitest/browser-playwright": "4.1.2",
         "@vitest/ui": "4.1.2",
         "cross-env": "10.1.0",
-        "esbuild": "0.27.4",
+        "esbuild": "0.28.0",
         "eslint": "10.1.0",
         "eslint-plugin-prettier": "5.5.5",
         "globals": "17.4.0",
@@ -70,6 +70,7 @@
         "tsup": "8.5.1",
         "typescript": "6.0.2",
         "typescript-eslint": "8.58.0",
+        "vite": "8.0.3",
         "vitest": "4.1.2"
     }
 }
