+++ src/test/unit/local-save.instance.test.ts
@@ -5,7 +5,7 @@ import {
     debugLog,
     isDebugLogsEnabled,
     randomString,
-} from '@local-save/test/test-utils';
+} from '@local-save/test/setup';
 
 describe('LocalSave - Instance', { tags: ['instance'] }, ({ beforeEach, afterEach }) => {
     beforeEach(({ task: { fullTestName } }) => {
